api/requirements.txt
flask
requests
gunicorn
promptflow
promptflow-core
promptflow-tracing
flask-cors
python-dotenv
azure-search-documents==11.4.0
